Update core.py
Add a func `time_to_stable` to claculate how long it will take for a glacier to stable in terms of volume change under the climatology of baseline-climate file around a center year (ex. if it is 1985, then this means the average climate in 1970-2000).
Change the mu range from [-2000,2000] to [-1000,1000].
The tolerance of area difference  in func `find_mb_offset` is set to 1e-3. (This is related to Landsat-image resolution, which is used to generate glacier inventory. For example, one pixel is 30m*30m=900m2=0.0009km2)